Added data visualization feature, among others
The UI now takes the data path as a command line argument and takes the
prediction results as a dataframe. After running the predictions, the UI
automatically generates a figure with layered histograms of the predicted powers
and measured powers. Works on the TestModel, will test on the
GaussianWeightedAverage once the predict functions are updated.